[Flow][Transforms] Add dynamic dim capture support to `scf.for` (#16889)
A post-condition of the flow compilation pipeline is the absence of
`tensor.dim`. Instead, the host module is supposed to have values
representing the dynamic dimensions available to it where needed.

When lowering some networks through IREE, e.g. RNNs, it may occur that a
`scf.for` is generated operating on tensor values. If these tensors have
dynamic dimension, the flow pipeline is unable to cope with them as the
dynamic dimensions of these tensors is not captured in the `sc.for`.

This PR fixes that issue by doing something similar to
`--iree-flow-capture-dispatch-dynamic-dims` but for `scf.for`: The
dynamic dimensions of loop variables is captured for both the
initializers and the yield operands and added as loop variables.
Finally, the tensor results are also tied with the dimension results.

I opted to do the bold move of generalizing the
`--iree-flow-capture-dispatch-dynamic-dims` pass and renaming it to
`--iree-flow-capture-dynamic-dims` that now additionally does the
capturing for `scf.for`. My thought process was that its position in the
pipeline ought to also be ideal for any captures happening in the
`scf.for`.