[LLVMGPU] Allow sending expanded convolutions down mfma pipeline (#16917)
This infers the convolution dimension types of linalg ops to determine
tile sizes to use. This is the first in a series of convolution focused
PRs from the sdxl branch. Because this is going down the vector
distribution pipeline which currently is not on by default and is only
used for mfma, so landing these changes is relatively low-risk.