Disable LLVM optional deps. (#16942)
From a user report, our Windows builds have started picking up an extra
dep on libxml2, which is then pulling in other things. I think this
might have been a combination of either an LLVM change or something on
the buildbots causing some use-if-available checks to pass and pull
things in.

In any case, disabling the things we don't need/want and will see how it
goes.