[LLVMGPU] Wmma layout for LLVMGPU vector distribute pipeline (#16928)
This PR introduces WMMA layout on to LLVMGPU pipeline. The main changes
are actually not too big, surrounding introduction of the WMMA layouts,
data-duplicate to express duplication of data through modification of
the thread basis, and emitting WMMA intrinsic. Large portions of the
changes are generalizing the names of classes and variables to represent
that we are doing MMA in general and not mfma specific things since most
parts of the mfma layout work that we have done is reusable.

A todo that I plan to handle after this patch is to get layout for
16x16x16 with an FP16 accumulator since it has a weird requirement to
further interleave the output/C-matrix data since it still only does 8
elements per wmma instruction on C-Matrix, however we want to represent
it as 16 elements where index=0,2,..,14 holds the real value.