[CPU] Do not use pre-configured tile sizes as input vector sizes for x86  (#16692)
It is not reliable for some cases. E.g., the below snippet is extracted
from mobilebert_int8 model and the pre-configured tile sizes of unpack
op is `[8, 4]`. The tile sizes are driven by consumers. We are able to
tile tensor.unpack op even the tile sizes are not aligned with
inner_tiles. The tradeoff is to generate a stack buffer to hold all data
required by tensor.unpack op. It is TilingInterface implementation
details. The op will always result in `*x8x8` shape, and the vector
input sizes are `*x8x8`. In this context, the configured tile sizes are
incorrect for vectorization. It is only correct for tiling.

```
%unpack = tensor.unpack %extracted_slice
  outer_dims_perm = [0, 1]
  inner_dims_pos = [0, 1]
  inner_tiles = [8, 8]
  into %17 : tensor<1x1x8x8xi32> -> tensor<8x8xi32>
%19 = linalg.generic ... { ... } -> tensor<8x4xi8>
%pack = tensor.pack %19
  outer_dims_perm = [0, 1]
  inner_dims_pos = [0, 1]
  inner_tiles = [8, 4]
  into %extracted_slice_2 : tensor<8x4xi8> -> tensor<1x1x8x4xi8>
```

The revision triggers an issue in quant_matmul_fusion pipeline_tests. It
is added a long while ago, and it's being outdated. All the matmuls on
x86 path are data-tiled. Thus, we retire the test in the revision.