Refresh "profiling with Tracy" developer docs. (#16939)
Progress on https://github.com/openxla/iree/issues/14470

|  |  |
| -- | -- |
| Current page |
https://iree.dev/developers/performance/profiling-with-tracy/ |
| This PR |
https://scotttodd.github.io/iree/developers/performance/profiling-with-tracy/
|

This restructures the "Profiling with Tracy" developer doc / website
page to focus less on the specifics of build and compile configurations
and more on what Tracy is and how to use it. Troubleshooting tips and
building from source instructions are still included but they are off to
the side. I also added new screenshots, diagrams, tips, and formatting.

Some of the "building from source" instructions can be further
simplified if we
* start using [Tracy's recently added upstream CMake
build](https://github.com/wolfpld/tracy/blob/master/CMakeLists.txt) (in
fact, the old Visual Studio solutions and Makefiles were removed in
https://github.com/wolfpld/tracy/commit/7f8a2e2c9a8581c2263da77222f1194ee0653e48)
* address https://github.com/openxla/iree/issues/16777