Allowing flow.tensor.constant to be used for constants. (#17024)
Prior it had only been used for dynamically shaped constants for
testing. Now it supports that mode (both for testing and programs that
want to ensure certain portions are treated dynamically, but mostly
testing) but also supports being used ala arith.constant. arith.constant
is still preferred for now as it composes with upstream stuff better but
our own op lets us support constant types arith cannot, such as
parameters. With this frontends can pass in IR with constant parameters
inline instead of needing to make them globals.

(I still don't really like that this op has the dynamic behavior and may
remove it in the future - the few uses of it can change to using casts
with optimization barriers on them as a more verbose but less nasty way
of doing the same thing)