[runtime] Add semaphore test where 2 batches wait on a former batch amongst 2 (#17080)
With this this test I added some functions to make the test body
shorter.
Unfortunately, we can't put gtest asserts in functions that have a
non-void return. Because of that in the functions I am using
`IREE_EXPECT*` instead of `IREE_ASSERT*`. This may cause tests other
than the one that is failing to fail also, as the whole executable may
crash if for example a command buffer fails to be created and then we
get a segmentation fault when trying to use it.
I thought this is an acceptable sacrifice to avoid repetition and to
have shorter and more readable tests.