[Codegen] Add folding in createBoundedTileSize for partially dynamic wgSize. (#17089)
With grouped-gemm op from dequant-gemm, when distirbuting using cyclic
distribution, we are still generating dynamic sizes. The cause of this
issue is the num_workgroup is partially dynamic, which in the current
pipeline would skip the affine min canonicalization completely. This PR
switches it to still attempt the min-workgroup-id fold if there are any
static dims found.

This is required to generate vector.contract which is required for MMA
operations in ROCM of dequant-gemm operations.