[CodeGen] Add a pattern to fold extract_slice consumer into xfer.write. (#17067)
The LLVMCPU and SPIRV lit tests are updated because the pass runs
`scf::ForOp` canonicalization patterns. It drops the unused for op
results which makes analysis easier.

This is helpful for bufferization when masking is not involved. Because
we make the chain simpler and the new tensor.empty op can be replaced
with `flow.tensor.load` ops in bufferization pre-processing passes.