[LLVMGPU] Modify layouts to be able to handle dequant operation. (#17113)
This PR solves 2 issues to enable dequant-gemm to work on vector
distribution:

1. During dequant-gemm, we will have reads of scale and zero-point that
will feed into a broadcast operation which will generate a vector layout
conflict operation. To resolve this, we do not anchor on these
transfer_read and depend on the layout from the transfer_read of quants.

2. We have not applied permutation on the subgroup and thread active IDs
which resulted in numerical issues during dequant-gemm. This PR applies
them and added a test for them.