Removing the `iree.compiler.consteval` attr. (#17056)
This tied the hoist-into-globals and jit-globals passes together in a
way that made it difficult to run them independently. Now the
jit-globals pass uses the const-expr analysis to determine if an
initializer can be evaluated without looking at attributes meaning that
new/combined initializers between when hoist-into-globals and
jit-globals run can be processed.

This required improving analysis to support tracking whether a
particular initializer only depends on other globals that are
initialized prior by the JIT pass. The analysis is fairly basic as it
just preserves the previous behavior but now correct instead of
supporting more things (like calls in initializers or control flow).
Whenever it gets reworked to be better it'd be useful to move out for
use in other places.

This causes some regressions in what we JIT (Falcon7bInt4GptqPT) due to
ConstExprAnalysis not being able to handle linalg ops performing
`iota`-like operations. It only worked by chance before (if the consumer
of the `iota`-like op _also_ consumed a constant root that was picked up
by analysis) and the proper fix is to rework the analysis to remove the
concept of constant roots and better identify what ops can be evaluated
at compile-time.