[CPU] Limit vectorization tile sizes (#17115)
The immediate motivation is fixing a long compilation time for a linalg
fusion that was taking 40 seconds to compile due to too-large
vectorization tile sizes, and takes 0.6 seconds with this PR:
https://gist.github.com/stellaraccident/83b357bbe2da31d5872cdbfcdd93aeea#file-example-mlir

However this addresses a more general issue: we've known that
vectorization tile sizes should be limited to what needs to fit within
register space vs. how big register space is, so this implements that on
some common CPU architectures where register space size is a
compile-time constant.