Add more unit tests for TileSizeSelection (#17219)
Adds unit tests for hooks that retrieve the actual level for the given
tiling level, e.g. `getDistributionLevel()`.

To improve code re-use, initialize of `LoweringConfig` is moved to a
dedicated hook, `initLoweringConfig`.

One death test is added to verify that an assertion is hit when
"out-of-bounds" tiling level is request.