Fixes to passes for custom dispatch to work with bf16 type (#17242)
Custom dispatch for bf16 type stopped working because of a couple of
problems:

1. `--iree-codegen-expand-strided-metadata` doesn't work with `iree-opt`
when given an `iree_codegen.extract_strided_metadata` op due to a lack
of a dependency on the `memref` dialect.
2. `--iree-convert-bf16-to-uint16-buffers` doesn't convert types in an
`iree_codegen.extract_strided_metadata` op due to a lack of a dependency
on the `iree_codegen` dialect. As a result, type casts are introduced
between the `iree_codegen.extract_strided_metadata` and its
`hal.interface.binding.subspan` predecessor, which confuses
`--iree-codegen-expand-strided-metadata`, which fails to do its job of
eliminating the `iree_codegen.extract_strided_metadata`. This second
problem was described in https://github.com/iree-org/iree/issues/17177.

The fix for both problems is to add the missing dialect dependencies.

fixes: https://github.com/iree-org/iree/issues/17177