[NFC] Refactoring MeshToFlow pass out from patterns. (#17245)
This moves the pass out to the Transforms/ directory and separates the
pass logic from the patterns for consistency with all the other
conversions. This will also help us move around where the pass is run
and possibly merge it with the main ConvertToFlow pass by separating
lookup from channel materialization as we do with device globals. Such
refactorings will (likely) be required for proper multi-device
scheduling but this initial refactoring is just moving code around and
leaves the pass running where it does (during the input pipeline).

We could probably put the ConvertMeshToFlow pass in
InputConversion/Common/ today based on where it's invoked, but I feel
like input may be the wrong layer for it if we expect any global
optimization phase work to happen on the mesh dialect and we may want to
move it later. It depends on whether we treat the flow ops as the
optimization mechanism for any partitioning work we do during global
optimization.