Converting some runtime benchmarks to use our C API. (#17336)
The API is incomplete and only works from C++ today but allows us to
hide the google benchmark dependency behind the iree::testing::benchmark
library. Some bazel/cmake goo can be used to switch off the google
benchmark library and no-op it (until we have an embedded-friendly impl)
but that's left as future work.

This should make solving what #16110 was doing much easier.