[CPU] Remove CPUDoubleTilingPeelingExpert (#17329)
`CPUDoubleTilingPeelingExpert` is effectively `CPUDoubleTilingExpert`
with the addition of loop peeling. This patch removes the former to
improve code re-use. In order to enable peeling, update the pipeline
config attribute to contain the `enable_loop_peeling` attribute:

   iree_codegen.translation_info<CPUDoubleTilingExpert, {enable_loop_peeling = true}>

This is a follow-up for https://github.com/iree-org/iree/pull/17231 in
which this new attribute was used to specialise
`CPUConvTileAndDecomposeExpert`.