[Winograd] Use output_tile_size for more static output transform tiling (#17200)
In the winograd output transform, the output tiles will always write in
bounds, and will always have the same static shape. This PR uses this in
the tiledImplementation of the output transform to maintain more static
information.

This also gets rid of an unnecessary alloc that comes from a dynamically
sized `tensor.empty` op after decomposition.