[Codegen] Add lane distribution for scf.forall (#17373)
This adds a new type of mapping id, iree_gpu.lane_id, and an op to
distribute forall ops mapping to it. Lane mapped forall ops require that
the thread count exactly matches with the subgroup count, so this is
added downstream due to incompatibility with upstream lowerings, however
this should be upstreamed in the future.