[GPU] Thread through a common target description (#17217)
This commit starts threading through a common target description
for various GPU backends. The goal is to unify how we specify
the target across CUDA, HIP, Vulkan, and others, so that we can
be consistent and reuse the some facilities.

Concretely, we introduce a `#iree_gpu.target` attribute in the
`IREEGPUDialect` to describe the GPU target. It includes a few
fields:

* The canonical target architecture for compilation, e.g.,
  sm_80 for cuda, gfx942 for hip
* A TargetWgpAttr describing the GPU features and limits in a
  single GPU workgroup processsor (that is, AMD compute unit
  or NVIDIA streaming multiprocessor), e.g., compute/storage
  capabilities, subgroup sizes, and workgroup size limits
* An optional TargetChipAttr describing GPU features for the
  final chip or product, e.g., wgp count

To help writing tests, we also introduce another command-line
option, `--iree-gpu-test-target` to specify the target. This option
will be queried if we cannot find the `#iree_gpu.target` attribute
in the wrapping IR. This avoids us to duplicate the full target
for each test to be concise and clear.

In order to support common development targets and the shorthand
attribute, we list some common AMD/NVIDIA GPU architectures
and their capabilities and limits. This is in general fine for
datacenter GPU use cases where we only ever care one architecture.

This starts with update both LLVMGPU backends (CUDA and ROCm)
to use the new target description. Vulkan side is untouched and
the next step to unify.

Progress towards https://github.com/iree-org/iree/issues/16341

ci-extra: test_nvidia_gpu,test_nvidia_a100,test_amd_mi250,test_amd_w7900