[Codegen][GPU] Add vectorization pattern for iree_gpu.multi_mma (#17453)
This allows vectorizing a statically shaped `iree_gpu.multi_mma` op. To
use this pattern with dynamically shaped tensors, a tiling
implementation for this op will need to be added in the future.

Additionally this adds a new set of `IREEGPUTransformExtensions` because
`Codegen/CommonExtensions` has become a rather large file and this is
specific to the dialect.