[Codegen][GPU] Allow iree_gpu.tensor_barrier to take vectors (#17479)
This allows synchronizing on vectors as well as tensors with similar
semantics. In a typical lowering flow, this will represent the
read equivalent to a tensor barrier, in that a tensor barrier represents
a wait until all writes to a shared allocation has finished, while this
represents a wait until all threads have read the value they need from
that shared allocation.

Renames the operation to iree_gpu.value_barrier for clarity.