[util] Add serialization support for `f64` resources (#17640)
Serializing `f64` resources was strangely omitted from the logic while
`f32` and `f16` support is present. Running things with `f64` is
certainly not a good idea in general but relatively well-supported in
the LLVM backend. Our use-case is the bring-up of a custom compiler
backend where `f64` happens to be the most trivial element type to
support.

Signed-off-by: Markus Böck <markus.boeck02@gmail.com>