[CPU] Signal errors if there are large vectors. (#17620)
The default ratio is 512 because the codegen could generate ops with
`vector<16x16x16xf32>` types in some tests. It should be revisited after
we move codegen to a better state.

An additional `iree-llvmcpu-fail-on-large-vector` flag is added. It is
on by default; it is a way to bypass the check for developers.

Additional changes:

- Update the preset tile sizes in `lowering_config.mlir`. It does not
matter because it is just testing the e2e flow for preset compilation.
- Disable the check for `sdxl-scheduled-unet-3-tank` model. It should be
removed after we fix the attention codegen issue.
- Disable the `sdxl-vae-decode-tank` model because of bad attention
codegen. Different from the other one, it is generating ops with
vector<512x32xf16> types, which should really just fail.

Fixes https://github.com/iree-org/iree/issues/17486

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>