Update CI to include benchmarking changes in Test Suite. (#17655)
This commit updates the SHARK-Test ref, config files, and yaml files to
have the most up to date flags and benchmarking support. I will also
concentrate on a python implementation for pulling in configs in
Test-Suite, so we don't have to rely on all these config files. Checked
the golden values over 15 times in the CI so should be good.

This commit also adds support to the CI, so that it generates a job
summary of the benchmark mean times for e2e and all the sub models. This
can be seen by developers in the summary tab of the PckgCI testing.
Example: https://github.com/iree-org/iree/actions/runs/9501523985

<img width="646" alt="image"
src="https://github.com/iree-org/iree/assets/77521230/8c0e8732-64a9-4147-b596-64520b0622d6">


Side note: The build_test_all_bazel was failing the first couple times
and then passed. Seems to be unstable

---------

Signed-off-by: saienduri <saimanas.enduri@amd.com>